# nuxk-core/api/openapi.yaml
openapi: 3.1.0
info:
  title: nuxk-core agent API
  version: "1"
  description: >
    The contract between the nuxk-core agent (on the router, or the Pi test
    stand) and its clients: the lite web UI served by the agent itself and
    nuxk-controller on the Pi.

    Single-tenant. Every path except /healthz and /auth/* needs either
    `Authorization: Bearer <API_TOKEN>` (`/opt/etc/nuxk/nuxk.conf`; programs —
    the controller) or the `nuxk_agent` session cookie from POST /auth/login
    (a browser: the box's root login and password, as for SSH). A session's
    writes must come from the agent's own page (Origin is checked). A loopback
    caller is let in without either only when no token is configured.
    Responses are the resource itself; errors are `{"error":{"code","message"}}`
    plus the HTTP status. Request bodies are capped at 4 MiB.

    This file is checked against the agent's route table in CI
    (TestOpenAPIMatchesRoutes): an endpoint exists here if and only if the
    agent serves it. Planned endpoints are not listed until they exist.
  license: { name: MIT }
servers:
  - url: http://{host}:4141
    variables: { host: { default: 192.168.1.1 } }
security:
  - bearer: []
  - session: []
...
paths:
...
  /api/v1/logs:
    get:
      tags: [agent]
      operationId: logs
      summary: The agent's recent log (last 2000 entries kept in memory)
      parameters:
        - { name: after, in: query, description: only entries with a larger seq, schema: { type: integer, minimum: 0 } }
        - { name: limit, in: query, description: "at most this many, newest (default 200, max 2000)", schema: { type: integer, minimum: 1, maximum: 2000 } }
      responses:
        "200":
          description: entries, oldest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/LogEntry" } }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/logs/debug:
    get:
      tags: [agent]
      operationId: logsDebug
      summary: The debug switch of the in-memory log
      responses:
        "200":
          description: the switch
          content: { application/json: { schema: { $ref: "#/components/schemas/LogDebug" } } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [agent]
      operationId: setLogsDebug
      summary: Debug logging on for a while (it turns itself off), or off
      description: >
        While on, debug entries — engine script calls, RCI requests, commands,
        DNS failures — reach the in-memory log (/logs, /events). The log file on
        the router's flash keeps info and up. The agent started with -debug is
        on for good ("forced").
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/LogDebugSet" } } }
      responses:
        "200":
          description: the switch after the change
          content: { application/json: { schema: { $ref: "#/components/schemas/LogDebug" } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/dns/log:
    get:
      tags: [dns]
      operationId: dnsLog
      summary: The last questions and their answers, newest first
      parameters:
        - { name: q, in: query, schema: { type: string }, description: part of the name }
        - { name: n, in: query, schema: { type: integer, minimum: 1, maximum: 200 }, description: "how many (200 by default)" }
      responses:
        "200":
          description: the questions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/DNSQuery" } }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    session: { type: apiKey, in: cookie, name: nuxk_agent }
...
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...
    LogEntry:
      type: object
      required: [seq, ts, level, msg]
      properties:
        seq: { type: integer }
        ts: { type: integer, description: unix milliseconds }
        level: { type: string, enum: [debug, info, warn, error] }
        msg: { type: string }
        attrs: { type: string }
    LogDebug:
      type: object
      required: [on]
      properties:
        on: { type: boolean }
        until: { type: integer, format: int64, description: unix milliseconds when it turns itself off }
        forced: { type: boolean, description: "started with -debug: on for the process' life" }
    LogDebugSet:
      type: object
      required: [on]
      properties:
        on: { type: boolean }
        minutes: { type: integer, minimum: 1, maximum: 240, description: how long (default 30) }
...
    DNSQuery:
      type: object
      required: [at, domain, type, source, ms]
      properties:
        at: { type: integer }
        domain: { type: string }
        type: { type: string, description: "A, AAAA, HTTPS…" }
        answer: { type: string, description: "the first addresses, or nxdomain / empty / rcode N" }
        source: { type: string, enum: [cache, upstream, stale, failed] }
        ms: { type: integer }
...